refactor: reuse `fctx.Check*ScopesPresence` in gRPC interceptors
