+++ context/helpers.go
@@ -1,16 +1,15 @@
-package foundation
+package context
 
 import (
 	"context"
 
-	fctx "github.com/ri-nat/foundation/context"
 	ferr "github.com/ri-nat/foundation/errors"
 )
 
 // CheckAllScopesPresence checks if the context contains all specified scopes.
 // If any of the specified scopes is missing, it returns an "insufficient scope" error.
 func CheckAllScopesPresence(ctx context.Context, scopes ...string) *ferr.PermissionDeniedError {
-	if !fctx.GetScopes(ctx).ContainsAll(scopes...) {
+	if !GetScopes(ctx).ContainsAll(scopes...) {
 		return ferr.NewInsufficientScopeAllError(scopes...)
 	}
 
@@ -20,7 +19,7 @@ func CheckAllScopesPresence(ctx context.Context, scopes ...string) *ferr.Permiss
 // CheckAnyScopePresence checks if the context contains at least one of the specified scopes.
 // If none of the specified scopes are present, it returns an "insufficient scope" error.
 func CheckAnyScopePresence(ctx context.Context, scopes ...string) *ferr.PermissionDeniedError {
-	if !fctx.GetScopes(ctx).ContainsAny(scopes...) {
+	if !GetScopes(ctx).ContainsAny(scopes...) {
 		return ferr.NewInsufficientScopeAnyError(scopes...)
 	}
