+++ examples/clubchat/chats/cmd/api/list_chats.go
@@ -3,7 +3,7 @@ package main
 import (
 	"context"
 
-	f "github.com/ri-nat/foundation"
+	fctx "github.com/ri-nat/foundation/context"
 	ferr "github.com/ri-nat/foundation/errors"
 	pb "github.com/ri-nat/foundation/examples/clubchat/protos/chats"
 )
@@ -17,7 +17,7 @@ func (s *chatsServer) ListChats(ctx context.Context, req *pb.ListChatsRequest) (
 	}
 
 	// Check required scopes
-	if err := f.CheckAllScopesPresence(ctx, ScopeProfile, ScopeChats); err != nil {
+	if err := fctx.CheckAllScopesPresence(ctx, ScopeProfile, ScopeChats); err != nil {
 		return nil, err
 	}
